% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_mapping.R
\name{add_grid_mapping}
\alias{add_grid_mapping}
\title{Add grid mapping}
\usage{
add_grid_mapping(x, mapping, overwrite = FALSE)
}
\arguments{
\item{x}{file name}

\item{mapping}{list of name of variable and attributes for mapping params (see DEtails)}

\item{overwrite}{enforce user override to actually update the file}
}
\value{
the name of the file modified, returned invisibly
}
\description{
grid_mapping by CF convention
}
\details{
mapping is a list with elements 'name' of the crs variable, and 'atts' list
with named attributes and values (e.g. standard_parallel = c(10, 20))
}
\examples{
mapping <- list(name = "crs",
                   atts = list(grid_mapping_name = "lambert_conformal_conic",
                               standard_parallel = c(-10, -50),
                               #standard_parallel = -50,
                               latitude_of_projection_origin = -30,
                               longitude_of_central_meridian =  134.33,
                               semi_major_axis = 6370000))

}
