% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wince-package.R
\docType{package}
\name{wince-package}
\alias{wince}
\alias{wince-package}
\title{wince: Write NetCDF Files for the Obvious Situations}
\description{
Create a 3D 'NetCDF' with time or depth as the third dimension, and convention-conforming coordinate system metadata.
}
\author{
\strong{Maintainer}: Michael D. Sumner \email{mdsumner@gmail.com}

}
\keyword{internal}
