% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_nc.R
\name{write_nc}
\alias{write_nc}
\title{Create raster file in NetCDF.}
\usage{
write_nc(
  data,
  filename = NULL,
  extent = NULL,
  title = "raster",
  zvar = NULL,
  z_type = NULL,
  data_name = NULL,
  data_unit = "some.unit",
  long_name = "some.long.name",
  params = "",
  overwrite = FALSE
)
}
\arguments{
\item{filename}{the NetCDF file to create}

\item{title}{name of the model}

\item{zvar}{actual time steps, need to be regularly space}

\item{overwrite}{set to \code{TRUE} to clobber an existing file}

\item{transp_params}{optional details to put in the NetCDF notes}
}
\value{
the filename of the output (use ncdf4 to inspect, modify it)
}
\description{
Rasters need extent, dimension, projection - this writer gives the correct metadata for simple
raster inputs with minimal generality and no fuss.
}
\details{
These functions
aim to minimize the amount of manual handling of details, creating an NetCDF file that can be modified directly.
}
\examples{
png <- system.file("textures/world.png", package = "rgl", mustWork = TRUE)
arr <- aperm(png::readPNG(png), c(2, 1, 3))
arr <- arr[,ncol(arr):1, ]
f <- write_nc(arr, extent = c(-180, 180, -90, 90), data_name = "world_image")
# terra::plotRGB(terra::rast(f) * 256)
# maps::map(add = T)
}
